backend/sources/requirements.txt
APScheduler==3.9.1
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
blinker==1.5
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
exceptiongroup==1.0.0rc9
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
install==1.3.5
kaitaistruct==0.10
lxml==4.9.1
numpy==1.23.3
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pip-review==1.2.0
pkg_resources==0.0.0
psutil==5.9.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
selenium==4.5.0
selenium-wire==5.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
urllib3-secure-extra==0.1.0
webdriver-manager==3.8.3
wsproto==1.2.0
zstandard==0.18.0
